rework TOC support to use XSL transformation via XmlPatterns
This fixes #1509 and was broken since upgrade to QT 4.8.5. The XSLT support
in Webkit was disabled in https://gitorious.org/webkit/qtwebkit/commit/8961d12
due to QTBUG-10309 and was imported in QT 4.8.0.

The XSLT support in XmlPatterns is still broken, however it should work for
the default stylesheet we use for the TOC generation. Any major changes
in the stylesheet may break/not work as expected due to the incomplete XSLT
support provided by XmlPatterns.

UPGRADE NOTE: the default TOC stylesheet has changed, so you will have to
redo any customizations based on the default stylesheet (the namespace
has changed to http://wkhtmltopdf.org/outline and it now uses XSLT 2.0).