Wait for an event before rendering the document
If the `--event` option is set, the rendering of the document
must wait for the event before triggering the action  If the
`--event-target` option is set, then this target is used rather
than `document`.

It is based on the windowStatus mechanism, and erases it ; when
the event is triggered, it changes the windowStatus (and also its
setting).