install debian-archive-keyring before creating the chroot environments
This is required on Ubuntu and validates that the Release signatures
are valid, and is already installed by default on Debian. Unfortunately,
there is no way we can do the same for Ubuntu as ubuntu-archive-keyring
is not yet present in the Debian stable release and the name differs
from the Ubuntu version (ubuntu-keyring vs ubuntu-archive-keyring).