perform the POSIX and OS X builds with -no-rpath
Using `-rpath` can be a security issue as it attempts to load the
libraries from `/home/build/wkhtmltopdf` which is unlikely to be
present on the target machine and can be potentially exploited
by an attacker. This can be mitigated by running `chrpath -d`
on the binaries, but is better to build without rpath -- it also
fixes #1733.