generate native packages for the target linux distribution using "fpm"
For the Debian-based distributions, we use XZ compression which results
in packages which have the same size as the tarballs which were generated
earlier. Unfortunately, there is no support for using XZ compression in
CentOS 5/6 and hence we have to use bzip2 which increases the file sizes
quite a bit (tarballs were around 16MiB while the package is around 42MiB).

All the native packages should have correct dependency information, but it
would unadvisable to install a package on a different distribution as only
the package names are specifed without specifying any version numbers.