refactor the OS X build system to use latest libjpeg/libpng libraries
Earlier, the bundled versions of the above libraries were used, which is
not safe as there might be security vulnerabilities in those versions.
The framework introduced in 9db4dce10d8ba81cb854d6994a842b6c9d149726 to
support building dependent libraries for Windows was reused to download
and compile the latest versions of the above libraries.

In addition, the build system was refactored to use "make install" support
which was introduced in bd16273f806759d82c9f64abaa2c093a76442dad for Linux.