generate native PKG for OS X using "fpm"
On OS X, packages are very limited and do not have the functionality that is
provided by package managers on Linux/BSD systems -- installation is supported,
but no uninstallation. Also, the compression support for packages is very poor
-- a 92MiB package is generated instead of a 14MiB tarball compressed using xz.

Thus, we use a very ugly hack to get the package down to a reasonable size. We
actually create a compressed tarball using xz (as before) and include a after-
install script which decompresses this tarball into /usr/local. To support the
scenario where the end-user might not have xz installed on the machine, a copy
of unxz is also bundled in the package.