download/compile xz for building/decompressing tarball in OS X installer
On OS X, we previously packed a locally installed xz into the installation
package and unpacked it during installation, if xz is not available on the
target machine. This can lead to problems if xz was installed by a package
manager like Homebrew, as symlinks would get packed instead of the real
binary.

We now download and compile xz statically on the fly, which removes the
need for xz on both the developer and the deployment machine.