set page margins correctly via header/footer heights for multiple URLs
In 80cbcc5db95eb48127a2d1418653e8643ad74f46, the ability for
auto-calculation of the header/footer height was implemented.
However, in case of multiple page objects (i.e. URLs) being present
for conversion, the calculated height for only the FIRST page object
was used -- the heights for the second and further page objects were
being ignored. In case a cover page was specified first (which has
no header/footer by design) without an explicit margin, then the
top/bottom margins would be set to zero -- which would hide the
header/footer content.

The solution is to use the maximum value across all the page objects,
so that there is enough space for the header content across all page
objects. Fixes #1676