block access to local files on Windows when an invalid URL is used
If you use an absolute path e.g. C:\TEST.PNG it is incorrectly
interpreted as an URL with the scheme as "C". This shows up as a
warning which was added in QTBUG-17731:

QNetworkAccessFileBackendFactory: URL has no schema set, use file:// for files

However, we should simply not process such URLs similar to what was
done for Qt5 via qtproject/qtbase@9581c90395086932d2e853864bf962b0896597d5.

Fixes #1639