fix handling of QList<X> in the reflection subsytem
assuming that <prefix> is an item which supports QList<X>
e.g. "load.cookies" then the following actions are supported:

<prefix>.length   get the length of the list ("count" and "size" also work)
<prefix>.first    get the first element or blank if list is empty
<prefix>.last     get the last element  or blank if list is empty
<prefix>[index]   gets the element specified by index

For setting the values, you can do the following:

<prefix>.clear    clears the list, value is ignored
<prefix>.append   appends  a blank item to the list, value is ignored
<prefix>.prepend  prepends a blank item to the list, value is ignored
<prefix>.delete   deletes an item (list index is specified via the value)
<prefix>[index]   sets the element specified by index to the given value

Error handling was also improved to return a failure when setting a value
or returning a blank string when trying to get a value. Fixes #1944