fix handling of temporary files during PDF conversion via API
The TempFile class was able to handle only one file, but was used
in scenarios where multiple temp files were required. Change the
implementation and API to support this usage scenario, which also
fixes #1790.

Ideally, one should do away with this and use an in-memory QtWebKit
protocol [1] e.g. mem://uuid.html but that is something to be taken
up after refactoring of the current codebase.

[1] http://doc.qt.digia.com/qq/32/qq32-webkit-protocols.html