fixes #2142
To be quite honest, I'm really not aware of any possible side-effects of
turning the loader into a mainLoader.

But it does fix #2142. wkhtmltoimage now respects `--javascript-delay`
and `--window-status` commandline flags.

So far, in all our tests on debian-linux, wkhtmltoimage works and
behaves as expected when the MultiPageLoader is configured as a
mainLoader as by this patch.