fix broken debug builds with MSVC
This requires special treatment as MSVC has a separate runtime
library which needs to be specified when compiling each source
file. This has to be done for all dependent libraries so that
there is no mismatch of the CRT used for compiling the complete
source code for dependent libraries, Qt and the application.

fixes #2322