create chroot environment for generic linux binaries
It is based on CentOS 6, without installing any of the following
libraries: libjpeg, libpng or libicu. These will be compiled and
linked statically using the deplibs functionality.

The remaining libraries are having a stable ABI (i.e. SONAME) and
can be used as-is across distributions with dynamic linkage. Using
CentOS 6 means that glibc will also not be a problem, as it uses
the oldest version we would prefer to support.