fix incorrect version detection when building distro-specific packages
This was broken since fe2e77c3b9dc5845c802ff096c1f6e45fe979c96,
as fpm is now called in the host environment. As the version is
available only for .deb packages and those will be deprecated in
the upcoming release, just record a dependency on the package.

Closes #2572