Set library install name on OS X
Libraries on OS X are referenced by binaries using absolute paths. When
linking, the linker copies this path from the library it links into the
compiled binary. To make this work with non-default paths, the library
must contain its own installation path as so-called "install name".

This makes libwkhtmltox a well-behaved library on OS X and fixes running
wkhtmltopdf if the library isn't in /usr/lib or /usr/local/lib.