use same QWebPrinter instance in "counting" and "printing" phases
There could be differences in layout during both these phases due
to the underlying WebCore::PrintContext forcing print layout in
the constructor and attempting to restore it back to screen mode
in the destructor. This did not always complete correctly, and in
the last phase it could result in less pages (blank page at end)
or more pages (crash when headers/footers were defined).

This should also speed up processing, as re-layout across phases
is now avoided.