Update createRequest warning message to indicate a possible solution
This change updates the `createRequest` signal on a `disposed`
`NetworkAccessManager` warning message to indicate one more issue that
may be fixed by a user controlled setting. This was discovered by
running with `--log-level debug` which I added in b8f9f5e354b2 /
https://github.com/wkhtmltopdf/wkhtmltopdf/pull/4668 and seeing the
debug message "Loading done; Stopping QWebPage and any possible page
refreshes." before I expected it to be called (before an API request had
finished). This is called by the `loadFinished` handler
https://github.com/wkhtmltopdf/wkhtmltopdf/blob/5b26acc41d8209d3b4c4018ac7350b8cea64fa73/src/lib/multipageloader.cc#L336-L340
and from that I realized that `loadDone` was being called too early. I
don't believe the immediate `loadDone()` is being called because the
default JavaScript delay is 200ms, but setting the timeout to 5 seconds
fixed the issue for me, and logging this suggestion as a warning may
help someone else when debugging.