Parse href in findLinks with QUrl::fromPercentEncoding
This allows for both internal and external links containing percent escaped characters to be parsed correctly.
Previously, http://example.com/foo%26bar was parsed as http://example.com/foo%2526bar instead of
http://example.com/foo%26bar (http://example.com/foo&bar). Same issue for other escaped characters.

According to the docs for the QUrl constructor, in it's default parsing using TolerantMode, it'll only
transform %20 back to space, and will treat all other % signs as raw symbols to be encoded.